domain refactored out; tests need to be updated
