+++ quarkus-cafe-web/pom.xml
@@ -5,7 +5,7 @@
   <parent>
         <groupId>com.redhat.quarkus</groupId>
         <artifactId>quarkus-cafe-demo</artifactId>
-        <version>1.0-SNAPSHOT</version>
+        <version>1.1-SNAPSHOT</version>
   </parent>
   <groupId>com.redhat.quarkus.cafe</groupId>
   <artifactId>quarkus-cafe-web</artifactId>
@@ -39,6 +39,13 @@
     </dependencies>
   </dependencyManagement>
   <dependencies>
+    <!-- Application domain -->
+    <!-- Application Domain -->
+    <dependency>
+      <groupId>com.redhat.quarkus</groupId>
+      <artifactId>quarkus-cafe-domain</artifactId>
+      <version>${project.version}</version>
+    </dependency>
     <dependency>
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-resteasy-jsonb</artifactId>
