--- quarkus-cafe-web/src/main/java/com/redhat/quarkus/cafe/web/domain/EventType.java
@@ -1,9 +1,0 @@
-package com.redhat.quarkus.cafe.web.domain;
-
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-@RegisterForReflection
-public enum EventType {
-
-    BEVERAGE_ORDER_IN, BEVERAGE_ORDER_UP, KITCHEN_ORDER_IN, KITCHEN_ORDER_UP
-}
