--- quarkus-cafe-web/src/main/java/com/redhat/quarkus/cafe/web/domain/Order.java
@@ -1,19 +1,0 @@
-package com.redhat.quarkus.cafe.web.domain;
-
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-@RegisterForReflection
-public class Order {
-
-    public Item item;
-
-    public String name;
-
-    public Order(Item item, String name) {
-        this.item = item;
-        this.name = name;
-    }
-
-    public Order() {
-    }
-}
