--- quarkus-cafe-web/src/main/java/com/redhat/quarkus/cafe/web/domain/OrderStatus.java
@@ -1,9 +1,0 @@
-package com.redhat.quarkus.cafe.web.domain;
-
-import io.quarkus.runtime.annotations.RegisterForReflection;
-
-@RegisterForReflection
-public enum OrderStatus {
-
-    IN_QUEUE, READY;
-}
