+++ quarkus-cafe-web/src/main/java/com/redhat/quarkus/cafe/web/infrastructure/JsonUtil.java
@@ -1,7 +1,5 @@
 package com.redhat.quarkus.cafe.web.infrastructure;
 
-import com.redhat.quarkus.cafe.web.domain.DashboardUpdate;
-import com.redhat.quarkus.cafe.web.domain.OrderEvent;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 
 import javax.json.bind.Jsonb;
@@ -14,19 +12,6 @@ public class JsonUtil {
 
     static String toJson(final Object object) {
         return jsonb.toJson(object);
-    }
-
-    static OrderEvent orderEventFromJson(final String string) {
-        return jsonb.fromJson(string, OrderEvent.class);
-    }
-
-    static String convertOrderEventToDashboardUpdate(final OrderEvent orderEvent) {
-        return toJson(new DashboardUpdate(orderEvent));
-    }
-
-    static String convertOrderEventToDashboardUpdate(final String payload) {
-        final OrderEvent orderEvent = orderEventFromJson(payload);
-        return toJson(new DashboardUpdate(orderEvent));
     }
 
 }
