+++ quarkus-cafe-web/src/main/java/com/redhat/quarkus/cafe/web/infrastructure/OrderService.java
@@ -1,7 +1,7 @@
 package com.redhat.quarkus.cafe.web.infrastructure;
 
 
-import com.redhat.quarkus.cafe.web.domain.CreateOrderCommand;
+import com.redhat.quarkus.cafe.domain.CreateOrderCommand;
 import io.quarkus.runtime.annotations.RegisterForReflection;
 import org.eclipse.microprofile.reactive.messaging.Channel;
 import org.eclipse.microprofile.reactive.messaging.Emitter;
