+++ quarkus-cafe-web/src/test/java/com/redhat/quarkus/cafe/web/infrastructure/KafkaResourceIT.java
@@ -1,13 +1,14 @@
 package com.redhat.quarkus.cafe.web.infrastructure;
 
-import com.redhat.quarkus.cafe.web.domain.*;
+import com.redhat.quarkus.cafe.domain.EventType;
+import com.redhat.quarkus.cafe.domain.Item;
+import com.redhat.quarkus.cafe.domain.OrderInEvent;
 import io.quarkus.test.junit.QuarkusTest;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import javax.inject.Inject;
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.WebTarget;
@@ -51,10 +52,10 @@ public class KafkaResourceIT extends BaseTestContainersIT{
         msgEventSource.open();
         assertTrue(msgEventSource.isOpen());
 
-        OrderEvent orderEvent = new OrderEvent(
-                UUID.randomUUID().toString(),
-                UUID.randomUUID().toString(),
+        OrderInEvent orderEvent = new OrderInEvent(
                 EventType.BEVERAGE_ORDER_IN,
+                UUID.randomUUID().toString(),
+                UUID.randomUUID().toString(),
                 "Minnie",
                 Item.CAPPUCCINO);
