--- quarkus-cafe-web/src/test/java/com/redhat/quarkus/cafe/web/infrastructure/NativeApiResourceIT.java
@@ -1,9 +1,0 @@
-package com.redhat.quarkus.cafe.web.infrastructure;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeApiResourceIT extends ApiResourceTest {
-
-    // Execute the same tests but in native mode.
-}
